android delete method param fix

+++ android/src/main/java/lt/imas/react_native_signal/RNSignalClientModule.java
@@ -269,7 +269,7 @@ public class RNSignalClientModule extends ReactContextBaseJavaModule {
     }
 
     @ReactMethod
-    public void deleteSignalMessage(String username, long timestamp, final Promise promise) {
+    public void deleteSignalMessage(String username, int timestamp, final Promise promise) {
         signalClient.deleteSignalMessage(username, timestamp);
         promise.resolve("ok");
     }
